flatpak: Create flatpak manifest
Flatpak is a method of building and distributing sandboxed
applications that is becoming increasingly important within
the GNU/Linux communities. It is useful for shipping
applications across distributions as well. This commit adds
support for making flatpaks of KeepassGtk.

GNOME Builder is a useful application that can handle building
and exporting flatpak bundles.