Add convenience features
It is now possible to save the last used unlock method to save time when unlocking next time.
Besides that it's now also possible to remember the last used keyfile when using composite unlocking.
Both features can be enabled and disabled in the settings. The first one is enabled by default.
This fixes #15