CredentialsIntegrityError is no more in pykeepass 3.2.1
It was split into ``CredentialsError``, ``HeaderChecksumError``,
or ``PayloadChecksumError``. Allow to use both old and new scheme for now
as the new exceptions are not really used in any other way.

The backwards compatability hack is quite local and should be removed once
we rely on version 3.2.1 of pykeepass (unfortunately, there is no programmatic
way to check the pykeepass version).