on_password_entry_changed: simplify logic and make level_bar error proof
This started as error_proofing the password_strength level bar which can
now return None on failure which we could not cope with.

However, the logic in this callback function was so horribly convoluted
that I had to simplify it while touching the function. It tested all
possible combinations, that it still works as before.