Fix Regression from libhandy-1 transformation
When we would use the mobile width (ie narrow windows),
the pathbar is supposed to show up at the bottom in the
Actionbar. For some reason, this does not work anymore, and
compained about a Viewport in a Viewport, so we can simplify
the glade object quite a bit. And past this commit the pathbar
does show up again when the window gets narrow.