Make Pathbar appear in all cases
Simplify the container in container logic. Remove those IFs, and always
move the pathbar. Remove the call to action_bar(), it was called twice
and thus we tried to add the pathbar twice to the header which caused
failure.

Relly the action_bar function should be a callback that is invoked when
the window size switched between mobile and desktop width.