Fix up pathbar
The function action_bar is called when the pathbar is added/removed
between the top headerbar and the bottom actionbar. However, it is not
redundancy safe, and when we hit a "pathbar button" it would try to
re-add the pathbar to the header even if it already was there.
Make that function robust by checking if it is already there and doing
nothing in that case. This is better but still not well-designed.

The correct thing would be to have a listener that is invoked when the
window width changes, eg by using a HdySqueezer that sends a signal
when it is "squeezed" and not all content fits in anymore.

WIth this commit, we do not get those GTK errors anymore that occur
when clicking on pathbar buttons.