pathbar: Rework query_page_update
Schedule more pages for destruction when we have finished editing a page.
Always discard the edited page, and if we modified it, even discard the
parent group page as it needs to be rebuild (probably).

This fixes for instance the issue that the password visibility was never
updated for all entries that we had openend in the past when we changed
the global setting for password visibility.

Simplify the logic of what to discard and to be honest, we do not want to
have old edit pages hanging around in memory in case we need them later.

Fixes: https://gitlab.gnome.org/World/PasswordSafe/-/issues/222