Remove logging_manager in favor of plain logging
We were passing around a lot of logging_managers that had to be stored as
instances in all kind of classes. This is not needed, and we can simply use
logging.debug(), logging.info() etc to log using the configured logger.

Remove logging_manager and convert to using plain old logging.