properties_dialog: delete buggy attempt to close dialog window on db lock
The deleted lines were an attempt to automatically close the entry property
dialog when a database is locked (e.g. through timeout). However it was
buggy and would not have worked. It performed self.properties.close() when
it meant self.properties_dialog.close(). (it would have crashed had it
actually tried because UnlockedDatabase has no attribute properties.)

Delete this attempt, we cannot manually lock a database while the modal
dialog is open, and when autolocking a database I think it is OK to keep
a modal dialog open.