main_window: set application during __init__
Rather than setting MainWindow.application from the outside post __init__
we set it directly during __init__, so it is available during the
construction. Also a intializater of MainWindow needs to care and know less
about the internals of the MainWindow.

Pylint complained about the potential use of NotImplemented.* here (which
could occur if the user forgets to set the application from the outside),
this commit would fix that.