CI: Switch to to debian bullseye image for the static analysis stage
The debian pygobject package is built for the python version handled
by the debian version. Therefore, it is not possible to use the
python:3.9 image, which is baed on debian buster and handles the
python 3.6 and 3.7 versions of pygobject.

This issue is solved by using the debian:bulseye image. This image
contains a recent version of pylint. Consequently, it does not need to
be installed via pip3.