unlocked_database: Correctly delete an entry with the menu
When an entry with the submenu the parent group is deleted instead of
the entry itself. This results in a crash.

The issue is fixed by deleting the entry and using the parent_group to
update the pathbar. Also, `show_page_of_new_directory` does not need
to be called two times.

Related: https://gitlab.gnome.org/World/PasswordSafe/-/issues/177