database_manager: Remove self.password
Outch, we set self.password when we create a DatabaseManager, but
when we changed the password we only set self.db.password, so
self.password became obsolete and wrong. It was accessed from the
database_settings_dialog.py to change passwords, so comparing to
obsolete passwords would have led to bugs.

In some parts of the code we have therefore a pattern of

  database_manager.set_database_password( "newpassword" )
  database_manager.password = "newpassword"

which is useless and prone to errors.

Make password a simple property that fetches self.db.password and
the setter function part of the property too. All occurrences of
atabase_manager.set_database_password() simply become
database_manager.password="moo"

This fixes actual bugs that could manifest itself (at least!) unexpectedly.