entry_page: Correctly replace attachment if the file already exists
When downloading an attachment, a new file is saved to the
disk. Gio.File.create is used to save the file. However, sometimes,
the filename already exists. In that case, Gio.File.create will fail
because it does not handle existing files.

This issue is fixed by using Gio.File.replace. This function creates a
new file if it does not exist and can overwrite an existing one.