main_window: Properly listen to window resizes
the "check-resize" event is used to update the main window and its
content for reponsive ui purposes. However, this event occurs every
time the window checks it has been resized. It does mean the window
has really been resized. Therefore, a lot of unecessary change_layout
calls occur.

Instead, the size-allocate vfunc can be used. This function is the
event handler of the resizing event.