unlocked_database: Do not remove non-existent element from the stack
When an entry is entered with the search, it sometimes belong to a
group which has not been displayed yet. If this entity is then
deleted or duplicated, the parent group is then displayed. The logic
to display the parent group relies on a two step process: remove the
group view from the stack and then create it and add it again to the
stack to force a full redraw of the view: the duplicated entry
becomes visible of the deleted entity is removed from the group view.
However, if the group has not been displayed yet, a non-existent
element tries to be removed from the stack. This results in a crah.

The issue is fixed by checking that the stack_page exists before
deleting it.

Closes: https://gitlab.gnome.org/World/PasswordSafe/-/issues/177