unlocked_database: Improve switch_stack_page usage
Switch stack page is always used when the current_group is
set. Therefore, it makes sense to set the current group in
switch_stack_page.

Also, a new page is created if it does not exist already. Therefore,
there is no need to check if the page exists before calling
switch_stack_page.