passphrase_generator: Use "-" as default passphrase separator
Rather than not defining anything, provide a default which would
allow us to skip the parameter if we are happy with the default.

Make "-" the default separator in our gsettings.schema

If the default separator of a passphrase is "", we simply glue
a list of words together which defies the purpose of a passphrase.
Use "-" as default separator.