passphrase_generator: Encapsulate in a class
Rather than simply using a global function, we encapsulate the
generate() phassphrase function in a class method of used to create
class Passphrase. Much easier to import too, than the previous:

from passwordsafe.passphrase_generator import generate as generate_passphrase

This gives us a Passphrase object which I plan to make a special case of a
Password class (to be created) which can contain lots of methods that are
specific to passwords (e.g. compare()'ing 2 passwords)