unlocked_database: fix up get_uuid() function
The opaque line get_parent().get_parent().get_parent().get_parent()
was actually trying to grab a GroupRow object (which was not obvious
from THAT code). So we do need to use get_uuid() instead of accessing
.uuid after all.

While fixing that, make the function a bit prettier by documenting things
and walking up the object hierarchy until we have the GroupRow object.
This makes the code a bit more robust in case the object
hierarchy should ever change, and it is a bit clearer what the code does.

Fixes: https://gitlab.gnome.org/World/PasswordSafe/-/merge_requests/139
Related: https://gitlab.gnome.org/World/PasswordSafe/-/merge_requests/139