icon.py: error proof get_icon()
icon.value can be `None` in the data, and we sometimes blindly look up
via get_icon(). Make the function robust by handling None input and also
do range checks (the keepass safe file could hand us back any number).

This way we do not have to range check in the code each time
we use this function. The function now returns the default icon in case of
invalid input.

Also, there is no need to construct a dict with number->icon_name and look
up the dict by number when all we want to do is access the list by index.
This function is called once per entry in a potentially long list of entries
to show.

Also rename it to get_icon_name() to distinguish better from get_icon() in
database_manager.