main_windows: sanitize create_new_database_instance.
create_new_database_instance is called in a separate thread to ... create
a new safe file. We passed it a "tab_title" as parameter and it fetched the
filepath to use from self.filechooser_dialog.
The tab title is exclusively derived from the filepath, so there is no need
to pass it in. On the other hand, there *could* be a race condition that
self.filechooser_dialog has become unaccessible by the time our thread
accesses it. So pass in the filepath directly and be done with it. Less
internal mucking of self.XXX attributes across threads.

DO NOT assign the string "liufhre86ewoiwejmrcu8owe" as default password
to every new safe file!!!

Finally, rename the create_new_database_instance to create_new_database which is
shorter and just as descriptive while touching its API anyway.