create_database: Copy stock database not in our UI thread
We would copy the stock database in our filechooser gui thread and then
create the DatabaseManager in a separate thread. Move the stock database
copying in that thread too. The sequence and purpose of "copy_database",
"create_new_database_instance" and "start_database_creation_routines"
were pretty non-intuitive.
Also, with passing the filepath around as parameters, there is no need
at all to keep the filechooser dialog hanging around as an attribute of
our main_window ALL the time. We can make it a local thing which goes out
of scope when it is not needed anymore.