fix crash when clicking on recently opened file entry
A recent change in assemble_first_start_screen() accidentally put in the
file's path instead of its URI into the label name. However on opening
we would expect an URI, so pass one it. This way we are safer in allowing
remote URIs, so that is a good change anyway.

Rename the horribly named
  assemble_first_start_screen() -> display_recent_files_list()
as it is shown each time we close the last opened database too, for instance.

Fixes: https://gitlab.gnome.org/World/PasswordSafe/-/issues/220
Related: Commit 5d235626