performance: Ditch unneeded entry_has_name()
We already performed an entry.get_name() and stored that. So lateron we do
not need to query if it has a username. Either it has or it is None and we
know that already. This is called once per entry in the group to be displayed
and this call took 5 seconds on my laptop for a simple db with 1000 entries.