entry|group_row: Don't wrap selection checkboxes in GtkBoxes
Constructing these rows is performance critical and a large part of the
time is spent in fetching and constructing those gtk widgets. So simplifying
those rows is a performance critical path.

Don't wrap and store the selection checkboxes in yet another GtkBox that we
explicitely hide when we are not in search mode. Ditch that additional
layer and hide the checkboxes directly. Actually we hide them in the glade
ui now and only need to explicitely show them when we construct the EntryRow
in selection mode.

This shaves off a bit of time loading the 1000 entry big.kdbx and has no
other side effects (so far :-)).