entry_row: Only fudge with icon classes if color is set
Give all entry_row icons a DarkIcon class by default in the glade ui
and only remove it and add class "BrightIcon" if a color has been set.

This saves us from dynamically fudging with most widgets after their construction.
It does not save loads of time, but a little.