unlocked_database: ditch get_entries_in_folder(self.current_element.uuid)
get_entries_in_folder(self.current_element.uuid) was passing in a UUID,
looking up the group, and returning the group's.entries. This happens everytime
we create a new stack, ie display a group, so it is worth optimizing.

As this is THE ONLY consumer of this function, ditch it and simply use
group.entries directly.