unlocked_database: Ditch get_groups_in_folder()
When creating a new stack, ie displaying a group we would do a
get_groups_in_folder(current_element.uuid) which would look up the current
group, ie current_element and return its subgroups.

As this was THE ONLY consumer of that function, drop it and directly do
current_group.subgroups. Also there is no need to treat the root group
any different from other groups, so we can drop the special treatment too.
As that is THE ONLY consumer of get_groups_in_root(), drop it too.