Put EntryRow's glade data in separate .ui file, makes creating EntryRow MUCH faster
It turns out the big culprit is that we add_resources using the HUGE unlocked_database.ui
for each label. It contains lots of stuff which we do not want to create a 1000 times.
Just putting the entry_row stuff in its own dedicated resource files and
adding that a 1000 times is MUCH faster.

Also do NOT compress the .ui file in our gresources file, it small and used often, so
uncompressing this thing cost us more than it brings. (we should examine if the same
holds true for the other .ui files there)

This commit makes it go from around 25 seconds to <3.5 seconds for creating 1000 entries
in the gui.