entry_page: Don't ask for the icon using an entries UUID
when we have the real entry instance at hand, so we do not need to look it up
again. Also, all those if icon == number lines can be "elif" as there is only
one condition that can be true and we do not need to evaluate them all.
Putting the common case of 0 as the first value to be tested.