entry_row: Use a gesture to handle the click event
An EntryRow contains a button (for its icon) which prevents opening
the enry when clicking on it. Indeed, the button intercepts the click
event. This issue can be fixed by using a GtkGesture and setting its
propagation phase to capture.
Besides, the "button-press-event" signal does not exist in
Gtk4.Gtk.Gesture need to be used instead. This is a first step for a
Gtk4 port.

Closes: #226