Fix empty shown bottom action bar while safe file loads
We were showing the bottom revealer by default and only hiding it after the
database file has been loaded. This causes aa brief visual glitch and depending
on the loading time it shows a black line at the bottom while loading.

Setting the revealer to "revealed" false by default and use the responsive_ui.
action_bar() function to perform the correct placement of the pathbar depending
on which mode we are in right from the beginning.

Fixes: #229