entry_row: Remove the button used to display the icon
The icon of the EntryRow is displayed by using a Image inside a
Button. The button is styled with the "circular" class to have a
circular icon. However, the button prevents the click events on the
EntryRow from being properly handled.

This is issue is fixed by removing the button. The same style can be
achieved by using a css effect with a 'border-radius' set to the same
size as the image.

This is based on an initial idea by Maximiliano Sandoval R.