entry_row: Switch to a bubble propagation phase for the controller
The capture propagation_phase was used to prevent the click events
from the button of the icons. However, the button to copy the entry
does not work because it is inside the widget of the controller. So,
the copy event is never reached.
With the icon button removed, it is now possible to use the bubble
propagation phase.