search: Do not perform a search if the search term is empty
A search is performed every time the search entry is changed. This
means that sometimes an empty search term is performerd. For example
if searching for '1' and then '0', the following search is performed
'1', then '' and finally '0'.
This can result in some strange effects as multiple search results are
displayed at the same time.

This issue is fixe by checking that the search entry is not empty
before performing a search.

Related: #224