search: Rename result_list to results_to_show in instance_creation
The result_list paramater is not the list of all the results from a
search but only the results that need to be shown. Indeed, when there
are a lot of results, some results are skipped by default and addded
to the skipped_rows list.
Then, these skipped results can be show by clicking on the 'show all
results' button. In that case, the search_instance_creation method is
called again with the skipped_rows as parameter.

This renaming helps avoiding any ambiguity.