search: Wait before performing a search
Usually, when peforming a search, multiple keys are entered by the
user before performing a search. However, in Search, a search is
performed everytime a new key is pressed. This can result in a lot of
unnecessary intermediate searchs and some strange effects as multiple
search results can be displayed at the same time.

This issue is fixed by adding a timeout. A search is only performed if
the entry has changed during 500ms.

Closes: #224