password_entry_row: Do not save the password if it is unchanged
The _on_password_value_changed callback is automatically called on
PasswordEntryRow construction. This results in a false change to the
database: the same password is saved to the database. Therefore, the
save dialog is displayed on exit even if there has been no change to
the database.

This issue is fixed by checking that the password is really different
before saving it to the database.

This is based on an idea by Maximiliano Sandoval R.

Closes: #237