unlocked_database: Do reset the headerbar after an element deletion
Since, the rewrite of selection_mode, linkedbox_right is not removed
from the headerbar when entering selection-mode and re-added when
leaving selection-mode. Only its visibility is changed.

When an entry or a group is deleted, the parent group view is
completely repopulated by a call to swith_page. Then, the main
headerbar becomes visible by calling set_browser_headerbar. However,
in selection-mode, the headerbar does not need to be updated.

Fix the issue by not calling set_browser_headerbar in selection-mode.