Fix title display in headerbar
Show the title only in mobile view or when no selection mode is active.
We achieve this simply by always keeping the title_label in our headerbar
(instead of the fragile addition and removal of the thing) and simply set the
visibility if we want to show/hide it.

It turned out that multiple classes/files were directly acessing the
filename_label and fundging with it (even if it was removed at the moment),
but were then all calling headerbar_title in responsive_ui.py anyway where
we can set the title. Remove those direct callers which were superfluous and
dangeroud. Title display works now again.

Fixes: #242