database_manager: Only use entry or group for get_parent_group
The application deals with Entry and Group and sometimes use their
uuid to get an entry or group. When an uuid is used, a lookup to the
database is needed to retrieve the corresponding Entry or Group. This
is quite inefficient as the Entry or Group was always known in the
first place.