fix self.is_dirty -> page.is_dirty typo
pylint discovered that we were setting self.is_dirty outside of __init__
and indeed that was a real mistake in the code (introduced by me). It was
supposed to set page.is_dirty (which is a thing) instead of self.is_dirty
(which does not exist for Pathbar())
This is more than a pylint beautification, it is a real bug discovered with
pylint's help.