application.py: disable pylint arguments-differ false positives
The parameters for do_startup, do_command_line, and do_activate are exactly
as they are (just look at the tutorials), but pylint somehow complains as it
thinks some (*args, **kwargs) should also be handed it or so (that is my theory).
Silence pylint here, we want these parameters to be exactly as they are.