Define safe_entry
SafeEntry is a keepass entry embedded into a GObject. It will make it
possible to simplify the code in the next commits. For example, it
will be possible for an EntryRow to be automatically updated by
binding the related properties to listening to property changes.

It will be used for the first time in the next commit.