unlocked_datbase: Remove useless else on while look without break
the else: branch will ALWAYS be executed when there is no "break" inside
the "while" branch, so remove "else:". While doing that sleep a little
in that endless loop, otherwise we keep our little CPU busy and warm while
it is loading something in the main thread.

credits go to pylint for finding this.