search: Compute the height of the list box
The result list is not always correclty displayed: an empty list is
sometimes displayed. Then, the view is correctly displayed once the
view is resized. Forcing a check_resize of ScrolledPage helps mitigate
the issue (see the previous commit). By adding, a height-request on
the list box, this issue is fixed in all cases.

group_row_height and entry_row_height are also updated to take into
account the margins.

Closes: https://gitlab.gnome.org/World/PasswordSafe/-/issues/247