main_window: fix & simplify invoke_initial_screen()
We were using os.path.exists() to test if filepath exists. However, the
recent files are saved as Gio.File.uris, so could contain things like
sftp://tmp/recent.kdbx which could be interesting with os.path.exist().
Make use of the proper Gio.File.query_exists() function to test existence
to cope with those cases.

This was a big bunch of "if elif elif". The main change is to simply
make the function return if one of the three criteria is fulfilled and
fall through to the display_welcome_screen if nothing matches.