Handy Clamp: Remove maximum_size
libhandy in 1.0 has sane defaults for maximum_size and threshold, no
need to set these. Some {top,bottom}_margins were defined on the code, these are not
needed as defined on the CSS files.

Using this the entries are a bit smaller than on stable, a suprise to be
sure, but a welcome one.

Fixes: https://gitlab.gnome.org/World/PasswordSafe/-/issues/251