Do not cash when there is no first_start_grid
WHen we cannot open a db, we want to open the error bar. However, in some cases
there is no first_start_grid and we would crash (try opening a safe, and while
it asks for the password, open another safe and select a .pdf file to experience
the crash). At least don't crash for now, but also don't show the error bar.
This needs better fixing.