UnlockedDatabase:bind_accelerator(): don't pass in unlocked_database
Because this *is* an method on the instance of the unlocked_database
instance so passing it in as a parameter of this funtion is
redundant. Found by looking at why pylint complained about no usage of
"self." in an instance method.
Adapt all 3 users of this function and it makes us hand around less stuff.