AppWindow: Use HdyApplicationWindow
This is a step towards moving the entire UI to ui files.

To make use of HdyApplicationWindow, the way of handling the headerbar
had to be changed since its API is a little different than
GtkApplicationWindow, now it is a stack which just switches its visible child when needed.

In the future, these headerbars should go into a single
locked_headbar.{ui, py}, see https://gitlab.gnome.org/World/PasswordSafe/-/merge_requests/254