Do not import or use the "construct" module
We were using the construct module earlier as pykeepass used to raise a
construct.CheckSumError on corrupt files. However, they have since
changed their code and split this into the pykeepass Exceptions:
CredentialsError, PayloadChecksumError, HeaderChecksumError.

We rely on a version of pykeepass anyway that has that, so there is NO
use in importing and testing that Exception, which nicely does away with
our check and need for python3-construct during build time.