Enable us to open files passed as parameters
Use do_handle_local_options instead of the do_commandline in order to
parse our custom command line option (-d). We do not want to take over
command line handling copletely (as signified by the app_flag), because
then opening a file using the command line will not work.

This patch re-enables us to pass a filename to the start of gnome-password
that will be immediately opened.

Fixes: https://gitlab.gnome.org/World/PasswordSafe/-/issues/197