color_button: Directly inherit from a flowboxchild
When a widget is added is inserted into a FlowBox, a FlowBoxChild is
inserted if the widget is not a FlowBoxChild itself. By direclty
inheriting from a FlowBoxChild, it simplifies the code and allows to
use the "child-activated" signal from the FlowBox. The css also needs
to be adapted to have the same circular effect.