Do not use hardcoded python module paths
On regular systems we are fine with using the standard python search paths.
The hardcoded paths prevented me from running our .deb package created with
the CI (debian sid) on a Debian testing system as it hardcoded
/lib/python3.9/* in its search path, while gnome-password was loaded using
python3.8. This created a mess.

One thing that needs testing is if that also works on flatpak or if we need
those paths on flatpak