entry_row: Bind name property
By adding those bindings, there is no need to set the is_dirty value
of the associated ScrolledPage and to repopulate the associated on
each name change. The EntryRow is now automatically updated on every
name change.

Also, an italic css value is added. It helps bringing this binding
forward by removing the markup usage.