entry_row: Bind icon-name property
By changing "icon-name" to a Gobject property, it is possible to
automatically update an EntryRow when its icon is changed. This way,
there is no need to set the is_dirty value of the associated
ScrolledPage and to repopulate the associated on each icon
change. The EntryRow is now automatically updated on every icon
change.