safe_entry: Properly handle icons not displayed during edition
The keepass format can handle up to 69 icons. However, the Entry
edition page can only select amonge 10 of those icons. This means than
an entry can have an entry which is not visible in the EntryPage. This
had two issues:
- the default entry was selected by default in the EntryPage
- when an entry was edited, the entry icon was replaced by the default
icon ("0").

Thoses two issues have been somewhat solved in a previous commit by
removing all the icons not handled by SafeEntry. With the new
SafeEntry object, it is possible to add support for those icons
without the two issues previously mentionned. This support is added by
adding two mechanisms:
- An Icon class is introduced. It has a visible property. If set to
False, it means that the icon should not be displayed in EntryPage.
- an icon_handled property is added to SafeEntry. If set to False, it
means that the icon of the entry is not handled by PasswordSafe.

In EntryPage, an hidden RadioButton is added to the view if the icon
is not handled to prevent the default icon from being selected.
The next commit will reintroduce the icons which are not handled by
PasswordSafe.