Switch out Show/Hide and Copy password buttons
As elaborated in #192: Currently the order of the Hide/view and the copy
buttons is not consistent. In some views we have the Show/Hide action as
the secondary icon IN the entry field, in others next to it.

Make this more consistent and put the Show/Hide "eye" in the entry field
This seems also more in line with what other applications do.

Closes: https://gitlab.gnome.org/World/PasswordSafe/-/issues/192