create_database: fix password comparison when creating a safe
We (err, *I* :-)) previously removed the database_manager.set_password_check()
function which was only used to be able to compare 2 passwords. I removed
all of its users ... but one ... as it turns out. All credit goes to
@msandova for finding this oversight.

This is a minimally invasive fix suitable for backporting to our
stable 4.x branch.

Fixes: https://gitlab.gnome.org/World/PasswordSafe/-/issues/290