database_manager: Correctly delete an entry attachment
There are two ways to delete an attachment:
1. entry.delete_attachment(attachment)
2. database.delete_binary(attachment.id)

The first way only removes the link between the attachment and the
entry while the second one removes the link and removes the file (its
binary).

As far as PasswordSafe is concerned, removing an attachment also means
removing the binary from the database. This means that `delete_binary`
needs to be called.

Currently, `delete_entry_attachment` in `DatabaseManager` calls both
operations. The second one `delete_attachment` because the first call
has already removed the link between the entry and the
attachment. This is not visible because the operation happens in a
try\except logic.
In other words, the `entry.delete_attachment(attachment)` call is
useless and can be safely removed.