Entry Page: Turn into ScrolledWindow subclass
During this MR this will be turn into a Template, this commit is to make
the transition smoother. NOTE: The button which shows all rows does not
work in this commit. This is the content of the next commit.

Removed the Scrolled Window form UI file for the momment, will be added
back once this is turn into a template.