entry_page: Don't show OTP secret on old entries
When opening an existing entry only fields which have non-nil values
should be displayed. In this particular case, if there is a secret
stored we do not want to show it either, since the user does not need
this information once it is added to Password Safe.

In case the user wants to modify this value, it is still available after
expanding the entry page.