password levelbar: Improve CSS
Added a new offset so that each interval has a proper color. Also used
colors from Palette rather than hard coded ones. An epsilon = -0.01 was
added to the offset to ensure the color only changes when the segments change.

The border was removed to be inline with the Adwaita stylesheet.