unlocked_db: use only already existing group
The reason is the following. Let G  be a group, the each time we go from
G to one of its entries and back to G new SafeGroup wrapping G will be
created. This is wasteful and creates unnecessary callbacks, the safe
group representing the element in the pathbar does not coincide with the
one in current element.

This can be done thanks to subclassing the unlocked_db page.