window: Simplify opening filechooser
If the filechooser filter only allows for files with the KeePass mimetype, there is no point in
checking if the mime type is correct later. The only exception is if the
file does not have the application/x-keepass2 mimetype but has a .kdbx
extension. It would be best to remove this check, but it is there in
case the user system is ill-configured and the mimetype is not
applied correctly, this can happen if there are installations of other
KeePass managers in the system or a fault of the Linux distribution.

It is better to allow this files than to alienate ill-configured machines.