flatpak: Remove gvfs interface
gvfs can be used to read/create files in a virtual filesystem (e.g. Google
Drive GNOME integration) using
our current file chooser dialogs. The problem is
that pykeepass will fail on the save operation, and therefore we cannot
create such safes (we fail when setting the password), nor modify them. With no clear way to communicate
this to the user it is best to remove this functionality in flatpak.

See https://github.com/construct/construct/issues/941