element: Only set timeout if the entry can expire
The timeout is now only set if the entry can expire.
The _check_expiration function is called every time the expires or
expiry_date property is changed to ensure that the timeout is set if
needed or to notify if the expiration date is passed.

This also ensures that the timeout is removed once the expiration date
is passed.