db_manager: Add a method to change db credentials
A regular save won't do, we want to be absolutely sure that if the
operation fails, the fields of the safe won't be the incorrect ones.

I.e. setting the value dbm.password = new_password should not result in
the incorrect password being reported if the save operation fails.

handling all of this can be done outside of dbm, but it is harder to
track.