entry_page: Use widget actions
This solves the problem that multiple actions were added to the window
and the action could fire in an undesired circumstance.

It has the disadvantage that the app shortcuts can't be used, but it
also means that the shortcuts will only do something when the widget is
on the top level.